% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_data.R
\name{data_prep}
\alias{data_prep}
\title{Data prepartion}
\usage{
data_prep(
  data,
  group_name = "group",
  time_name = "time",
  id_name = "id",
  outcome_name = "Y",
  treat_name = "W",
  control_formula = "X",
  group_levels = c("a", "b")
)
}
\arguments{
\item{data}{The data}

\item{yname}{Name of outcome variable}

\item{tname}{Name of time variable}

\item{idname}{Name of id variable}

\item{gname}{Name of group variable}

\item{xformula}{Covariates to use in estimation of propensity score and outcome regression}
}
\description{
Data prepartion
}
