% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dgp.R
\name{example_dgp}
\alias{example_dgp}
\title{Example DGP}
\usage{
example_dgp(n, seed = NULL)
}
\arguments{
\item{n}{Number of units}

\item{seed}{The random seed}
}
\value{
A 2-by-2-by-2 dataset
}
\description{
The data generating process of the example
}
