% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{pred_rf}
\alias{pred_rf}
\title{predicitions based on a random forest models}
\usage{
pred_rf(mod, X)
}
\arguments{
\item{mod}{Random forest model}

\item{X}{Covariates (matrix)}
}
\value{
Predictions
}
\description{
predicitions based on a random forest models
}
