% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{train_rf}
\alias{train_rf}
\title{training random forest models}
\usage{
train_rf(Y, X, num.trees = 500, ...)
}
\arguments{
\item{Y}{Outcome (vector)}

\item{X}{Covariates (matrix)}

\item{num.trees}{Number of trees}

\item{...}{additional options for ranger}
}
\value{
A trained random forest model
}
\description{
training random forest models
}
