% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Data prepartion}
\usage{
data_prep(
  data,
  group_name = "group",
  target_group = "a",
  time_name = "time",
  id_name = "id",
  outcome_name = "Y",
  treat_name = "W",
  control_formula = "X"
)
}
\arguments{
\item{data}{The data}

\item{group_name}{Name of group variable}

\item{target_group}{Character / numeric value of the value in group_name that corresponds to the target group (group "a" in the paper)}

\item{time_name}{Name of time variable}

\item{id_name}{Name of id variable}

\item{outcome_name}{Name of outcome variable}

\item{control_formula}{Covariates to use in estimation of propensity score and outcome regression}
}
\description{
Data prepartion
}
