% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtdid.R
\name{drtdid}
\alias{drtdid}
\title{Double robust TDID estimation}
\usage{
drtdid(
  data,
  group_name = "group",
  time_name = "time",
  id_name = "id",
  outcome_name = "Y",
  treat_name = "W",
  control_formula = "X",
  type = "parametric",
  crossfit = FALSE
)
}
\arguments{
\item{data}{The data}

\item{group_name}{Name of group variable}

\item{time_name}{Name of time variable}

\item{id_name}{Name of id variable}

\item{outcome_name}{Name of outcome variable}

\item{treat_name}{Name of treatment variable}

\item{control_formula}{Covariates to use in estimation of propensity score and outcome regression}

\item{crossfit}{If true, estimation is done using 2-fold cross fitting}
}
\value{
DID and TDID estimates
}
\description{
PUT Desc here
}
